Added function to flush cache
Due to limitation of what seems to be everything in the toolchain, I
wasn't able to tell GCC or libtool or anything to compile a single
function as ARM and leave everything else as thumb AND use LTO. Since
LTO saves a decent amount of space, I deem that more important than
being able to seamlessly use both Thumb or ARM. As such, the library
must be compiled in the same mode as the final executable. I'll probably
provide an option for configure to do that soon.