Added experimental system clock support
-API BREAK: Changed some function names in ctr_timer. Moved some
functions from ctr_interrupt to the new ctr_irq.

-Implemented something that can be used as a system clock-- i.e. be used
to keep track of time. It relies on one of the 4 hardware timers, and it
needs the IRQ subsystem (also new) to be initialized. Haven't tested the
timespec-like poll function (higher resolution timing poll), but did
test the millisecond query.
-Implemented an IRQ subsystem where IRQ handlers can be registered.
ctr_system_clock uses this to register a timer IRQ.
-Forgot to backup include the original LR before an exception in
ctr_interrupt_asm.c . It is now included, useful for debugging.