GFX changes and interrupt tweaks
-API break: ctr_gfx object no longer exists-- all gfx operations now
happen or are handled at the screen level.

-Moved pointer for dealing with ITCM for interrupt setup to the first
mirror at 0x8000. This is so that programmers can set this second mirror
as r/w to set up the ITCM and set the base at 0x0 as executable only, so
invalid access to address 0x0 (which tends to be NULL) will outright fail
instead of actually going through. This is a concern the developer needs
to deal with in their MPU setup, and is something I cannot deal with at
the library level. Best I can do is accomodate it by moving access to
setup interrupts to some ITCM mirror not 0x0.