Added function to detect if SD card is inserted.
-Function added for ctr_sd_interface to determine if SD card is
inserted. Right now the function may not work for the early boot case,
since it seems super early boot the SD card flag that's being detected
is not very stable, or takes some time to trigger. Perhaps for the
future it may be better to track the state of the SD card by polling it
early until it is deemed "stable," then tracking changes via interrupts?

-Fixed an internal bug calculating the size of the SD card for fatfs
internal purposes.