Interrupts no longer require CPU mode stack setup
-Interrupt support no longer requires each individual CPU mode to have
its stack initialized to something. The behavior now is to use the SP
from the previously executing mode.
-Just in case the previously executing mode for interrupts was USER, a
check was put in place to prevent switching to USER mode from the
interrupt internal logic. Internal logic now switches to SYSTEM if USER
was the previous mode since SYSTEM shares all registers with USER and is
privileged.